+++ front/src/components/ui/resizable/ResizablePanelGroup.vue
@@ -1,6 +1,11 @@
 <script setup lang="ts">
 import { cn } from '@/lib/utils'
-import { SplitterGroup, type SplitterGroupEmits, type SplitterGroupProps, useForwardPropsEmits } from 'reka-ui'
+import {
+  SplitterGroup,
+  type SplitterGroupEmits,
+  type SplitterGroupProps,
+  useForwardPropsEmits,
+} from 'reka-ui'
 import { computed, type HTMLAttributes } from 'vue'
 
 const props = defineProps<SplitterGroupProps & { class?: HTMLAttributes['class'] }>()
@@ -15,7 +20,10 @@ const forwarded = useForwardPropsEmits(delegatedProps, emits)
 </script>
 
 <template>
-  <SplitterGroup v-bind="forwarded" :class="cn('flex h-full w-full data-[panel-group-direction=vertical]:flex-col', props.class)">
+  <SplitterGroup
+    v-bind="forwarded"
+    :class="cn('flex h-full w-full data-[panel-group-direction=vertical]:flex-col', props.class)"
+  >
     <slot />
   </SplitterGroup>
 </template>
