+++ front/src/components/ui/skeleton/Skeleton.vue
