+++ front/src/router/index.ts
@@ -2,8 +2,8 @@ import Auth from '@/views/Auth/index.vue'
 import Booking from '@/views/Client/Booking.vue'
 import Reservation from '@/views/Client/Reservation.vue'
 import Adm from '@/views/Manager/PaginaAdmTeste.vue'
-import Search from '@/views/Flights/Search.vue'
-import details from '@/views/Flights/Details.vue'
+import Search from '@/views/Client/Flights/Search.vue'
+import details from '@/views/Client/Flights/Details.vue'
 import BuyMiles from '@/views/Client/Miles/BuyMiles.vue'
 
 import { createRouter, createWebHistory } from 'vue-router'
