+++ front/src/views/Auth/index.vue
@@ -1,7 +1,21 @@
 <script setup lang="ts">
-import Auth from '@/components/features/Auth/Auth.vue'
+import { Tabs, TabsList, TabsTrigger, TabsContent } from '@/components/ui/tabs'
+import LoginForm from './components/LoginForm.vue'
+import RegisterForm from './components/RegisterForm.vue'
 </script>
 
 <template>
-  <Auth />
+  <Tabs default-value="login" class="w-[400px] mx-auto">
+    <TabsList class="grid grid-cols-2 w-full">
+      <TabsTrigger value="login">Login</TabsTrigger>
+      <TabsTrigger value="register">Cadastro</TabsTrigger>
+    </TabsList>
+
+    <TabsContent value="login">
+      <LoginForm />
+    </TabsContent>
+    <TabsContent value="register">
+      <RegisterForm />
+    </TabsContent>
+  </Tabs>
 </template>
