+++ front/src/views/Client/Flights/Details.vue
@@ -1,0 +1,10 @@
+<script setup lang="ts">
+import FlightDetails from './components/FlightDetails.vue'
+import FlightingPayment from './components/FlightingPayment.vue'
+</script>
+<template>
+  <div class="flex items-center min-h-screen justify-center">
+    <FlightDetails />
+    <FlightingPayment />
+  </div>
+</template>
