+++ front/src/views/Client/Flights/components/FlightCardsGrid.vue
@@ -11,7 +11,7 @@ defineProps<{
 <template>
   <div class="grid grid-cols-12 mt-4 gap-4 w-11/12 max-w-5xl">
     <template v-if="loading">
-      <Skeleton class="col-span-3 h-[229px] w-full rounded-xl" v-for="number in 8" />
+      <Skeleton class="col-span-3 h-[229px] w-full rounded-xl" v-for="number in 8" :key="number" />
     </template>
     <template v-else>
       <slot v-for="flight in flights" :flight="flight" />
