+++ front/src/views/Client/Flights/components/FlightDetails.vue
