+++ front/src/views/Client/Flights/components/FlightingPayment.vue
