+++ front/src/views/Client/Miles/components/MilesCheckout.vue
@@ -1,79 +1,69 @@
 <script setup lang="ts">
-import { computed } from 'vue';
-import { Button } from '../../../../components/ui/button';
-import {
-  FormControl,
-  FormField,
-  FormItem,
-  FormLabel,
-  FormMessage,
-} from '../../../../components/ui/form';
-import { Input } from '../../../../components/ui/input';
-import { Separator } from '../../../../components/ui/separator';
-import { useToast } from '../../../../components/ui/toast';
-import { useCheckoutForm } from '../composables/useCheckoutForm';
-import { useMilesPurchaseStore } from '../../../../stores/miles-purchase';
+import { computed } from 'vue'
+import { Button } from '@/components/ui/button'
+import { FormControl, FormField, FormItem, FormLabel, FormMessage } from '@/components/ui/form'
+import { Input } from '@/components/ui/input'
+import { Separator } from '@/components/ui/separator'
+import { useToast } from '@/components/ui/toast'
+import { useCheckoutForm } from '../composables/useCheckoutForm'
+import { useMilesPurchaseStore } from '@/stores/miles-purchase'
 
-const { toast } = useToast();
-const milesPurchaseStore = useMilesPurchaseStore();
+const { toast } = useToast()
+const milesPurchaseStore = useMilesPurchaseStore()
 
 const { handleSubmit, cardNumber, cardName, expiryDate, cvv } = useCheckoutForm()
 
 const formatCardNumber = (e: Event) => {
-  const target = e.target as HTMLInputElement;
-  let value = target.value.replace(/\s+/g, '');
+  const target = e.target as HTMLInputElement
+  let value = target.value.replace(/\s+/g, '')
   if (value.length > 0) {
-    value = value.match(new RegExp('.{1,4}', 'g'))!.join(' ');
+    value = value.match(new RegExp('.{1,4}', 'g'))!.join(' ')
   }
-  cardNumber.value.value = value;
-};
+  cardNumber.value.value = value
+}
 
 const formatExpiryDate = (e: Event) => {
-  const target = e.target as HTMLInputElement;
-  let value = target.value.replace(/\D/g, '');
-  
+  const target = e.target as HTMLInputElement
+  let value = target.value.replace(/\D/g, '')
+
   if (value.length <= 2 && target.value.includes('/')) {
-    value = value.substring(0, 2);
-    expiryDate.value.value = value;
-    return;
+    value = value.substring(0, 2)
+    expiryDate.value.value = value
+    return
   }
-  
+
   if (value.length >= 2) {
-    value = value.substring(0, 2) + '/' + value.substring(2, 4);
+    value = value.substring(0, 2) + '/' + value.substring(2, 4)
   }
-  
-  expiryDate.value.value = value;
-};
+
+  expiryDate.value.value = value
+}
 
 const onSubmit = handleSubmit(async (values) => {
-  console.log('Informações de pagamento:', values);
-  
+  console.log('Informações de pagamento:', values)
+
   toast({
     title: 'Pagamento efetuado com sucesso',
     description: 'Suas milhas foram compradas com sucesso!',
     variant: 'default',
-  });
-});
+  })
+})
 
-const totalPrice = computed(() => milesPurchaseStore.totalPrice);
+const totalPrice = computed(() => milesPurchaseStore.totalPrice)
 </script>
 
 <template>
   <div class="h-full w-full flex flex-col border-0 space-y-6">
     <div class="p-6">
-      <h2 class="text-lg font-semibold text-center">
-        Finalização de pagamento
-      </h2>
+      <h2 class="text-lg font-semibold text-center">Finalização de pagamento</h2>
     </div>
-    
+
     <div class="px-6 pb-6">
       <Separator />
     </div>
-    
+
     <div class="px-6 pb-6">
-      <h3 class="text-lg font-semibold text-center pb-4">
-        Detalhes do pagamento
-      </h3>
+      <h3 class="text-lg font-semibold text-center pb-4">Detalhes do pagamento</h3>
       <form @submit.prevent="onSubmit">
         <div class="space-y-4">
           <FormField v-slot="{ componentField }" name="cardNumber">
@@ -96,11 +86,7 @@ const totalPrice = computed(() => milesPurchaseStore.totalPrice);
             <FormItem>
               <FormLabel>Nome do titular</FormLabel>
               <FormControl>
-                <Input
-                  type="text"
-                  placeholder="Maria João"
-                  v-bind="componentField"
-                />
+                <Input type="text" placeholder="Maria João" v-bind="componentField" />
               </FormControl>
               <FormMessage>{{ cardName.errorMessage.value }}</FormMessage>
             </FormItem>
@@ -127,21 +113,14 @@ const totalPrice = computed(() => milesPurchaseStore.totalPrice);
               <FormItem>
                 <FormLabel>CVV</FormLabel>
                 <FormControl>
-                  <Input
-                    type="text"
-                    placeholder="123"
-                    v-bind="componentField"
-                    maxlength="4"
-                  />
+                  <Input type="text" placeholder="123" v-bind="componentField" maxlength="4" />
                 </FormControl>
                 <FormMessage>{{ cvv.errorMessage.value }}</FormMessage>
               </FormItem>
             </FormField>
           </div>
 
-          <Button type="submit" class="w-full mt-6">
-            Pagar R${{ totalPrice }},00
-          </Button>
+          <Button type="submit" class="w-full mt-6"> Pagar R${{ totalPrice }},00 </Button>
         </div>
       </form>
     </div>
