+++ front/src/views/Client/Miles/components/MilesQuantity.vue
@@ -1,41 +1,39 @@
 <script setup lang="ts">
-import { computed } from 'vue';
+import { computed } from 'vue'
 import { ref } from 'vue'
-import { Label } from '../../../../components/ui/label'
+import { Label } from '@/components/ui/label'
 import {
   NumberField,
   NumberFieldContent,
   NumberFieldDecrement,
   NumberFieldIncrement,
   NumberFieldInput,
-} from '../../../../components/ui/number-field'
-import { Separator } from '../../../../components/ui/separator'
-import { useAuthStore } from '../../../../stores/auth'
-import { Card, CardContent, CardHeader, CardTitle } from '../../../../components/ui/card';
-import { useMilesPurchaseStore } from '../../../../stores/miles-purchase'
+} from '@/components/ui/number-field'
+import { Separator } from '@/components/ui/separator'
+import { useAuthStore } from '@/stores/auth'
+import { Card, CardContent, CardHeader, CardTitle } from '@/components/ui/card'
+import { useMilesPurchaseStore } from '@/stores/miles-purchase'
 
 const authStore = useAuthStore()
 const milesPurchaseStore = useMilesPurchaseStore()
 
 const user = ref({
   firstName: authStore.user?.name.split(' ')[0],
-  miles: authStore.user?.miles
+  miles: authStore.user?.miles,
 })
 
 const miles = computed({
   get: () => milesPurchaseStore.miles,
-  set: (value: number) => milesPurchaseStore.setMiles(value)
+  set: (value: number) => milesPurchaseStore.setMiles(value),
 })
 </script>
 
 <template>
   <div class="h-full w-full flex flex-col border-0 space-y-6">
     <div class="p-6">
-      <h2 class="text-lg font-semibold text-center">
-        Informações de {{ user.firstName }}
-      </h2>
+      <h2 class="text-lg font-semibold text-center">Informações de {{ user.firstName }}</h2>
     </div>
-    
+
     <div class="px-6 pb-6">
       <ul class="space-y-4">
         <li>
@@ -44,15 +42,13 @@ const miles = computed({
         </li>
       </ul>
     </div>
-    
-    <Separator/>
-    
+
+    <Separator />
+
     <div class="p-6 pt-12">
-      <h2 class="text-lg font-semibold text-center">
-        Selecione a quantidade de milhas
-      </h2>
+      <h2 class="text-lg font-semibold text-center">Selecione a quantidade de milhas</h2>
     </div>
-    
+
     <div class="px-6 pb-6">
       <NumberField id="miles" v-model="miles" :min="0">
         <Label>Milhas</Label>
@@ -67,9 +63,7 @@ const miles = computed({
     <div class="px-6 pb-6">
       <Card>
         <CardHeader>
-          <CardTitle class="text-center">
-            Resumo da compra
-          </CardTitle>
+          <CardTitle class="text-center"> Resumo da compra </CardTitle>
         </CardHeader>
         <CardContent class="space-y-4">
           <div class="flex justify-between">
