--- front/src/views/Flights/Details.vue
@@ -1,10 +1,0 @@
-<script setup lang="ts">
-import FlightDetails from '@/components/features/Flights/components/FlightDetails.vue'
-import FlightingPayment from '@/components/features/Flights/components/FlightingPayment.vue'
-</script>
-<template>
-  <div class="flex items-center min-h-screen justify-center">
-    <FlightDetails />
-    <FlightingPayment />
-  </div>
-</template>
